divided old graph management component into subcomponents, added confirmation modal
